requirements.txt
flask==3.1.0
opencv-python-headless==4.11.0.86
pillow==11.1.0
numpy==2.2.4
werkzeug==3.1.3
pytesseract==0.3.13
python-dotenv==1.0.0